pyproject.toml
[project]
name = "maritime-segmentation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "onnx>=1.20.1",
    "onnxruntime>=1.23.2",
    "pycocotools>=2.0.11",
    "pyyaml>=6.0.3",
    "rfdetr>=1.4.0.post0",
    "supervision>=0.27.0",
    "tensorboard>=2.20.0",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "transformers>=4.57.3",
    "wandb>=0.23.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]
